+++ src/crash-course.sh
@@ -1,13 +1,17 @@
 #!/bin/bash
 
- read -p "What is this : " this
- echo "This is $this"
+read -p "What is this : " this
+echo "This is $this"
 
-echo $0
-echo $1
-echo $2
+# echo $0
+# echo $1
+# echo $2
 echo "${@}" # Access all the arguments [More on this later]
 
+for i in "$@"
+do
+  echo "$i"
+done
 
 echo "/////////////////////////:"
 
@@ -54,4 +58,70 @@ else
   echo "created by bash" | cat >> notes.md
 fi
 
+# 2. Arithmetic Evaluations
+read -p "Enter your age: " age
+
+if (( "$age" > 18 ))
+then
+  echo "Adult!"
+elif (( "$age" > 12 ))
+then
+  echo "Teen!"
+else
+  echo "Kid"
+fi
+
+
+# print numbers 1 to 10
+
+# c like for loop
+for (( i = 1; i <= 10; ++i ))
+do
+  echo "$i"
+done
+
+# for in
+for i in {1..10}
+do
+  echo "$i"
+done
+
+# while
+i=1
+while [[ "$i" -le 10 ]]
+do
+  echo "$i"
+  ((i++))
+done
+
+# until
+i=1
+until [[ "$i" -eq 11 ]]
+do
+  echo "unitl $i"
+  ((i++))
+done
+
+# iterating over arrays
+arr=(a b c d)
+
+# For in
+for i in "${arr[@]}"
+do
+  echo "$i"
+done
+
+# c like for
+for (( i = 0; i < "${#arr[@]}"; i++))
+do
+  echo "${arr[$i]}"
+done
+
+# while
+i=0
+while [[ "$i" -le "${#arr[@]}" ]]
+do
+  echo "${arr[$i]}"
+  (( i++ ))
+done