+++ CHANGELOG.md
@@ -6,6 +6,34 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.4.6] - 2021-05-11
+### Added
+- Using pass to handle passwords
+
+## [0.4.5] - 2021-04-16
+### Changed
+- Crash course: Iterating over arguments
+- Finetuning buscard
+
+## [0.4.4] - 2021-04-09
+### Changed
+- Crash course: Iterating over arrays
+
+## [0.4.3] - 2021-04-09
+### Changed
+- Update bus card
+
+## [0.4.2] - 2021-04-02
+### Added
+- Add a looping
+- Update bus card
+
+## [0.4.1] - 2021-03-26
+### Added
+- A CLI business card
+### Changed
+- Bash course conditionals
+
 ## [Released]
 
 ## [0.4.0] - 2021-03-19